Make bits types larger

+++ src/cpp/bits.hpp
@@ -121,7 +121,7 @@ template <class T> class BitsGeneric {
 
     // Converts from unit128_t to Bits. If the number of bits of value is smaller than size, adds 0 bits at the beginning.
     // i.e. Bits(5, 10) = 0000000101
-    BitsGeneric<T>(uint128_t value, uint16_t size) {
+    BitsGeneric<T>(uint128_t value, uint32_t size) {
         // TODO(mariano) remove
         if (size < 128 && value > ((uint128_t)1 << size)) {
             std::cout << "TOO BIG FOR BITS" << std::endl;
@@ -130,7 +130,7 @@ template <class T> class BitsGeneric {
         this->last_size_ = 0;
         if (size > 128) {
             // Get number of extra 0s added at the beginning.
-            uint16_t zeros = size - Util::GetSizeBits(value);
+            uint32_t zeros = size - Util::GetSizeBits(value);
             // Add a full group of 0s (length 128)
             while (zeros > 128) {
                 AppendValue(0, 128);
@@ -147,12 +147,12 @@ template <class T> class BitsGeneric {
 
     // Copy the content of another Bits object. If the size of the other Bits object is smaller
     // than 'size', adds 0 bits at the beginning.
-    BitsGeneric<T>(const BitsGeneric<T>& other, uint16_t size) {
-        uint16_t total_size = other.GetSize();
+    BitsGeneric<T>(const BitsGeneric<T>& other, uint32_t size) {
+        uint32_t total_size = other.GetSize();
         this->last_size_ = 0;
         assert(size >= total_size);
         // Add the extra 0 bits at the beginning.
-        uint16_t extra_space = size - total_size;
+        uint32_t extra_space = size - total_size;
         while (extra_space >= 128) {
             AppendValue(0, 128);
             extra_space -= 128;
@@ -161,14 +161,14 @@ template <class T> class BitsGeneric {
             AppendValue(0, extra_space);
         // Copy the Bits object element by element, and append it to the current Bits object.
         if (other.values_.size() > 0) {
-            for (uint8_t i = 0; i < other.values_.size() - 1; i++)
+            for (uint32_t i = 0; i < other.values_.size() - 1; i++)
                 AppendValue(other.values_[i], 128);
             AppendValue(other.values_[other.values_.size() - 1], other.last_size_);
         }
     }
 
     // Converts bytes to bits.
-    BitsGeneric<T>(uint8_t* big_endian_bytes, uint32_t num_bytes, uint16_t size_bits)  {
+    BitsGeneric<T>(uint8_t* big_endian_bytes, uint32_t num_bytes, uint32_t size_bits)  {
         this->last_size_ = 0;
         uint32_t extra_space = size_bits - num_bytes * 8;
         // Add the extra 0 bits at the beginning.
@@ -210,12 +210,12 @@ template <class T> class BitsGeneric {
         }
         BitsGeneric<T> result;
         if (values_.size() > 0) {
-            for (uint8_t i = 0; i < values_.size() - 1; i++)
+            for (uint32_t i = 0; i < values_.size() - 1; i++)
                 result.AppendValue(values_[i], 128);
             result.AppendValue(values_[values_.size() - 1], last_size_);
         }
         if (b.values_.size() > 0) {
-            for (uint8_t i = 0; i < b.values_.size() - 1; i++)
+            for (uint32_t i = 0; i < b.values_.size() - 1; i++)
                 result.AppendValue(b.values_[i], 128);
             result.AppendValue(b.values_[b.values_.size() - 1], b.last_size_);
         }
@@ -226,7 +226,7 @@ template <class T> class BitsGeneric {
     template <class T2>
     BitsGeneric<T>& operator += (const BitsGeneric<T2>& b) {
         if (b.values_.size() > 0) {
-             for (uint8_t i = 0; i < b.values_.size() - 1; i++)
+             for (uint32_t i = 0; i < b.values_.size() - 1; i++)
                 this->AppendValue(b.values_[i], 128);
             this->AppendValue(b.values_[b.values_.size() - 1], b.last_size_);
         }
@@ -244,14 +244,14 @@ template <class T> class BitsGeneric {
             bool all_one = true;
             if (values_.size() > 1) {
                 // Otherwise, search for the first bucket that isn't full of 1 bits.
-                for (int16_t i = values_.size() - 2; i >= 0; i--)
+                for (int32_t i = values_.size() - 2; i >= 0; i--)
                     if (values_[i] != limit) {
                         all_one = false;
                         // Increment it.
                         values_[i]++;
                         // Buckets that were full of 1 bits turn all to 0 bits.
                         // (i.e. 10011111 + 1 = 10100000)
-                        for (uint16_t j = i + 1; j < values_.size(); j++)
+                        for (uint32_t j = i + 1; j < values_.size(); j++)
                             values_[j] = 0;
                         break;
                     }
@@ -277,7 +277,7 @@ template <class T> class BitsGeneric {
         }
         if (values_.size() > 1) {
             // Search for the first bucket different than 0.
-            for (int16_t i = values_.size() - 2; i >= 0; i--)
+            for (int32_t i = values_.size() - 2; i >= 0; i--)
                 if (values_[i] != 0) {
                     all_zero = false;
                     // Decrement it.
@@ -286,7 +286,7 @@ template <class T> class BitsGeneric {
                                       (uint128_t)std::numeric_limits<uint64_t> :: max();
                     // All buckets that were previously 0, now become full of 1s.
                     // (i.e. 1010000 - 1 = 1001111)
-                    for (uint16_t j = i + 1; j < values_.size() - 1; j++)
+                    for (uint32_t j = i + 1; j < values_.size() - 1; j++)
                         values_[j] =  limit;
                     values_[values_.size() - 1] = (last_size_ == 128) ? limit :
                                                    ((static_cast<uint128_t>(1) << last_size_) - 1);
@@ -309,7 +309,7 @@ template <class T> class BitsGeneric {
         assert(GetSize() == other.GetSize());
         BitsGeneric<T> res;
         // Xoring individual bits is the same as xor-ing chunks of bits.
-        for (uint16_t i = 0; i < values_.size(); i++)
+        for (uint32_t i = 0; i < values_.size(); i++)
             res.values_.push_back(values_[i] ^ other.values_[i]);
         res.last_size_ = last_size_;
         return res;
@@ -361,7 +361,7 @@ template <class T> class BitsGeneric {
     }
 
     // Same as 'Slice', but result fits into an uint64_t. Used for memory optimization.
-    uint64_t SliceBitsToInt(int16_t start_index, int16_t end_index) const {
+    uint64_t SliceBitsToInt(uint32_t start_index, uint32_t end_index) const {
         /*if (end_index > GetSize()) {
             end_index = GetSize();
         }
@@ -392,7 +392,7 @@ template <class T> class BitsGeneric {
         // Append 0s to complete the last byte.
         uint8_t shift = Util::ByteAlign(last_size_) - last_size_;
         uint128_t val = values_[values_.size() - 1] << (shift);
-        uint16_t cnt = 0;
+        uint32_t cnt = 0;
         // Extract byte-by-byte from the last bucket.
         uint8_t iterations = last_size_ / 8;
         if (last_size_ % 8)
@@ -403,7 +403,7 @@ template <class T> class BitsGeneric {
         }
         // Extract the full buckets, byte by byte.
         if (values_.size() >= 2) {
-            for (int16_t i = values_.size() - 2; i >= 0; i--) {
+            for (int32_t i = values_.size() - 2; i >= 0; i--) {
                 uint128_t val = values_[i];
                 for (uint8_t j = 0; j < 16; j++) {
                     buffer[cnt++] = (val & 0xff);
@@ -412,7 +412,7 @@ template <class T> class BitsGeneric {
             }
         }
         // Since we extracted from end to beginning, bytes are in reversed order. Reverse everything.
-        uint16_t left = 0, right = cnt - 1;
+        uint32_t left = 0, right = cnt - 1;
         while (left < right) {
             std::swap(buffer[left], buffer[right]);
             left++;
@@ -422,11 +422,11 @@ template <class T> class BitsGeneric {
 
     std::string ToString() const {
         std::string str = "";
-        for (uint16_t i = 0; i < values_.size(); i++) {
+        for (uint32_t i = 0; i < values_.size(); i++) {
             uint128_t val = values_[i];
-            uint16_t size = (i == values_.size() - 1) ? last_size_ : 128;
+            uint8_t size = (i == values_.size() - 1) ? last_size_ : 128;
             std::string str_bucket = "";
-            for (int i = 0; i < size; i++) {
+            for (uint8_t i = 0; i < size; i++) {
                 if (val % 2)
                     str_bucket = "1" + str_bucket;
                 else
@@ -448,10 +448,10 @@ template <class T> class BitsGeneric {
         return values_[0];
     }
 
-    uint16_t GetSize() const {
+    uint32_t GetSize() const {
         if (values_.size() == 0) return 0;
         // Full buckets contain each 128 bits, last one contains only 'last_size_' bits.
-        return (values_.size() - 1) * 128 + last_size_;
+        return ((uint32_t)values_.size() - 1) * 128 + last_size_;
     }
 
     void AppendValue(uint128_t value, uint8_t length) {
@@ -520,7 +520,7 @@ bool operator==(const BitsGeneric<T>& lhs, const BitsGeneric<T>& rhs) {
     if (lhs.GetSize() != rhs.GetSize()) {
         return false;
     }
-    for (uint16_t i = 0; i < lhs.values_.size(); i++) {
+    for (uint32_t i = 0; i < lhs.values_.size(); i++) {
         if (lhs.values_[i] != rhs.values_[i]) {
             return false;
         }
@@ -532,7 +532,7 @@ template <class T>
 bool operator<(const BitsGeneric<T>& lhs, const BitsGeneric<T>& rhs) {
     if (lhs.GetSize() != rhs.GetSize())
         throw std::string("Different sizes!");
-    for (uint16_t i = 0; i < lhs.values_.size(); i++) {
+    for (uint32_t i = 0; i < lhs.values_.size(); i++) {
         if (lhs.values_[i] < rhs.values_[i])
             return true;
         if (lhs.values_[i] > rhs.values_[i])
@@ -545,7 +545,7 @@ template <class T>
 bool operator>(const BitsGeneric<T>& lhs, const BitsGeneric<T>& rhs) {
     if (lhs.GetSize() != rhs.GetSize())
         throw std::string("Different sizes!");
-    for (uint16_t i = 0; i < lhs.values_.size(); i++) {
+    for (uint32_t i = 0; i < lhs.values_.size(); i++) {
         if (lhs.values_[i] > rhs.values_[i])
             return true;
         if (lhs.values_[i] < rhs.values_[i])
@@ -592,7 +592,7 @@ BitsGeneric<T> operator>>(BitsGeneric<T> lhs, uint32_t shift_amount) {
     int num_blocks_shift = static_cast<int>(shift_amount / 128);
     uint32_t shift_remainder = shift_amount % 128;
 
-    for (int i = 0; i < lhs.values_.size(); i++) {
+    for (uint32_t i = 0; i < lhs.values_.size(); i++) {
         uint128_t new_value = 0;
         if (i - num_blocks_shift >= 0) {
             new_value += (lhs.values_[i - num_blocks_shift] >> shift_remainder);
